refactor: rename module folder to have valid python name
